requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
jinja2==3.1.2